% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_card.R
\docType{class}
\name{action_card}
\alias{action_card}
\title{Create a \code{action_card} Structure With \code{teamr}}
\format{
An \code{\link{R6Class}} generator object
}
\usage{
# ac <- action_card$new(type = "ActionCard", name = "Add comment")
}
\description{
\code{action_card} creates a \code{potentialAction} object which can be added to \code{connector_card}.
}
\details{
Assemble \code{action_card} objects into a \code{action_card}
structure and use the methods to append, modify or send your card to the webhook specified.
}
\section{Methods}{


\describe{
\item{\code{action_card$new(type, name)}}{Creates a new \code{action_card}.}
\item{\code{add_text_inputs(id, title, is_multiline)}}{Add \code{TextInput} to your \code{action_card}.}
\item{\code{add_date_inputs(id, title)}}{Add \code{DateInput} to your \code{action_card}.}
\item{\code{add_mchoice_inputs(id, title, is_multi_select)}}{Add \code{MultichoiceInput} to your \code{action_card}.}
\item{\code{dump()}}{Dump current payload}
}
}

\section{Properties}{


\describe{
\item{\code{payload}}{R list of payloads(will be parsed into json)\link[httr]{POST}}
}
}

\examples{
\dontrun{
library(teamr)

cc <- connector_card$new(hookurl = "https://outlook.office.com/webhook/...")
cc$text("Of on affixed civilly moments promise explain")
pa <- action_card$new(type = "ActionCard", name = "Add comment")
pa$add_actions()
pa$add_text_inputs(id = "comment", title = "Add comment for this task", is_multi_line = TRUE)
cc$add_potential_action(pa)
cc$send()
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-action_card-new}{\code{action_card$new()}}
\item \href{#method-action_card-add_actions}{\code{action_card$add_actions()}}
\item \href{#method-action_card-add_text_inputs}{\code{action_card$add_text_inputs()}}
\item \href{#method-action_card-add_date_inputs}{\code{action_card$add_date_inputs()}}
\item \href{#method-action_card-add_mchoice_inputs}{\code{action_card$add_mchoice_inputs()}}
\item \href{#method-action_card-dump}{\code{action_card$dump()}}
\item \href{#method-action_card-clone}{\code{action_card$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-new"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$new(type, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-add_actions"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-add_actions}{}}}
\subsection{Method \code{add_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$add_actions(
  type = "HttpPOST",
  name = "Save",
  target = "http://..."
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-add_text_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-add_text_inputs}{}}}
\subsection{Method \code{add_text_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$add_text_inputs(id, title, is_multi_line)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-add_date_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-add_date_inputs}{}}}
\subsection{Method \code{add_date_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$add_date_inputs(id, title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-add_mchoice_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-add_mchoice_inputs}{}}}
\subsection{Method \code{add_mchoice_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$add_mchoice_inputs(id, title, is_multi_select, choices)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-dump"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-dump}{}}}
\subsection{Method \code{dump()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-action_card-clone"></a>}}
\if{latex}{\out{\hypertarget{method-action_card-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{action_card$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
