% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_section.R
\docType{class}
\name{card_section}
\alias{card_section}
\title{Create a \code{card_section} Structure With \code{teamr}}
\format{
An \code{\link{R6Class}} generator object
}
\usage{
# sec <- card_section$new()
}
\description{
\code{card_section} creates an \code{Section} object which can be added to \code{connector_card}.
}
\details{
Assemble \code{card_section} objects into a \code{action_card}
structure and use the methods to append, modify or send your card to the webhook specified.
}
\section{Methods}{


\describe{
\item{\code{card_section$new()}}{Creates a new \code{card_section}.}
\item{\code{text(sec_text)}}{Change the \code{text} field of the section.}
\item{\code{title(sec_title)}}{Change the \code{title} field of the section.}
\item{\code{activity_title(sec_activity_title)}}{Change the \code{title} field of the section.}
\item{\code{activity_sub_title(sec_activity_subtitle)}}{Change the \code{subtitle} field of the section.}
\item{\code{activity_image(sec_activity_image)}}{Change the \code{image} field of the section.}
\item{\code{activity_text(sec_activitiy_text)}}{Change the \code{text} field of the section.}
\item{\code{add_fact(fname, fvalue)}}{Add \code{fact} object to section.}
\item{\code{add_link_button(btext, burl)}}{Add button with links.}
\item{\code{add_image(sec_image, sec_title)}}{Change webhook address.}
\item{\code{print()}}{Print out the current payload.}
\item{\code{dump()}}{Dump current payload.}
}
}

\section{Properties}{


\describe{
\item{\code{payload}}{R list of payloads(will be parsed into json)\link[httr]{POST}}
}
}

\examples{
\dontrun{
library(teamr)

cc <- connector_card$new(hookurl = "https://outlook.office.com/webhook/...")
cc$text("Of on affixed civilly moments promise explain")
sec$text("2018-19 Finals MVP")
sec$activity_sub_title("Kawhi Leonard")
sec$add_fact("Position", "Forward")
sec$title("Player Info")
sec$add_image(sec_image = "https://....jpg", sec_title = "so funny")
sec$activity_image("https://.....jpg")
sec$activity_title("Activity Title")
sec$activity_sub_title("Activity Subtitle")
sec$activity_text("Activity text")
cc$add_section(new_section = sec)
cc$print()
cc$send()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-card_section-new}{\code{card_section$new()}}
\item \href{#method-card_section-print}{\code{card_section$print()}}
\item \href{#method-card_section-title}{\code{card_section$title()}}
\item \href{#method-card_section-text}{\code{card_section$text()}}
\item \href{#method-card_section-activity_title}{\code{card_section$activity_title()}}
\item \href{#method-card_section-activity_sub_title}{\code{card_section$activity_sub_title()}}
\item \href{#method-card_section-activity_image}{\code{card_section$activity_image()}}
\item \href{#method-card_section-activity_text}{\code{card_section$activity_text()}}
\item \href{#method-card_section-add_fact}{\code{card_section$add_fact()}}
\item \href{#method-card_section-add_link_button}{\code{card_section$add_link_button()}}
\item \href{#method-card_section-add_image}{\code{card_section$add_image()}}
\item \href{#method-card_section-dump}{\code{card_section$dump()}}
\item \href{#method-card_section-clone}{\code{card_section$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-new"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-print"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-title"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-title}{}}}
\subsection{Method \code{title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$title(sec_title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-text"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-text}{}}}
\subsection{Method \code{text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$text(sec_text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-activity_title"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-activity_title}{}}}
\subsection{Method \code{activity_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$activity_title(sec_activity_title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-activity_sub_title"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-activity_sub_title}{}}}
\subsection{Method \code{activity_sub_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$activity_sub_title(sec_activitiy_subtitle)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-activity_image"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-activity_image}{}}}
\subsection{Method \code{activity_image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$activity_image(sec_activitiy_image)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-activity_text"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-activity_text}{}}}
\subsection{Method \code{activity_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$activity_text(sec_activitiy_text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-add_fact"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-add_fact}{}}}
\subsection{Method \code{add_fact()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$add_fact(fname, fvalue)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-add_link_button"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-add_link_button}{}}}
\subsection{Method \code{add_link_button()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$add_link_button(btext, burl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-add_image"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-add_image}{}}}
\subsection{Method \code{add_image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$add_image(sec_image, sec_title = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-dump"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-dump}{}}}
\subsection{Method \code{dump()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-card_section-clone"></a>}}
\if{latex}{\out{\hypertarget{method-card_section-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{card_section$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
